[BH-1515] Revert bedtime changes
Due to many regressions these changes
introduced, we decided to revert them.

The old PR will be used as a foundation for
another fix attempt(probably including major
refactor of the alarm operations module).
