+++ products/BellHybrid/alarms/include/popups/BedtimeReminderPopupRequestParams.hpp
@@ -1,20 +1,16 @@
-// Copyright (c) 2017-2022, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
 
 #include <apps-common/popups/data/PopupRequestParams.hpp>
-#include <Tables/Record.hpp>
 
 namespace gui
 {
     class BedtimeReminderPopupRequestParams : public PopupRequestParams
     {
       public:
-        BedtimeReminderPopupRequestParams(Record record)
-            : PopupRequestParams{gui::popup::ID::BedtimeNotification}, eventRecord{record}
+        BedtimeReminderPopupRequestParams() : PopupRequestParams{gui::popup::ID::BedtimeNotification}
         {}
-
-        Record eventRecord;
     };
 } // namespace gui
