+++ products/BellHybrid/alarms/src/actions/NotifyGUIAction.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #include "NotifyGUIAction.hpp"
@@ -10,7 +10,7 @@ namespace alarms
 {
     NotifyGUIAction::NotifyGUIAction(sys::Service &service) : service{service}
     {}
-    bool NotifyGUIAction::execute([[maybe_unused]] Record record)
+    bool NotifyGUIAction::execute()
     {
         return service.bus.sendUnicast(std::make_shared<app::actions::AlarmTriggeredAction>(), service::name::appmgr);
     }
