+++ products/BellHybrid/alarms/src/actions/NotifyGUIBedtimeReminderAction.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
@@ -13,7 +13,7 @@ namespace alarms
     {
       public:
         explicit NotifyGUIBedtimeReminderAction(sys::Service &service);
-        bool execute(Record record) override;
+        bool execute() override;
         bool turnOff() override;
 
       private:
