+++ products/BellHybrid/alarms/src/actions/PlayAudioActions.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
@@ -19,7 +19,7 @@ namespace alarms
     {
       public:
         bool turnOff() override;
-        bool execute(Record record) override;
+        bool execute() override;
         explicit PlayAudioAction(sys::Service &service,
                                  const std::filesystem::path &tonesDirPath,
                                  std::string_view toneSetting,
