+++ products/BellHybrid/apps/common/CMakeLists.txt
@@ -128,7 +128,6 @@ target_link_libraries(application-bell-common
         module-gui
         bell::db
         Microsoft.GSL::GSL
-        bell::alarms
         )
 
 if (${ENABLE_TESTS})
