+++ products/BellHybrid/apps/common/include/common/popups/BedtimeNotificationWindow.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
@@ -23,14 +23,11 @@ namespace gui
 
         app::ApplicationCommon *app;
         Icon *icon = nullptr;
-        Record currentEventRecord;
         bool onInput(const InputEvent &inputEvent) override;
         void returnToPreviousWindow();
         void buildInterface() override;
 
       public:
-        bool handleSwitchData(gui::SwitchData *data) override;
-
         explicit BedtimeNotificationWindow(app::ApplicationCommon *app);
         void onBeforeShow(ShowMode mode, SwitchData *data) override;
     };
