+++ products/BellHybrid/services/appmgr/include/appmgr/messages/AlarmMessage.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2022, Mudita Sp. z.o.o. All rights reserved.
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
@@ -6,7 +6,7 @@
 #include <Service/Message.hpp>
 #include <popups/AlarmActivatedPopupRequestParams.hpp>
 #include <popups/AlarmDeactivatedPopupRequestParams.hpp>
-#include <popups/data/BedtimeReminderPopupRequestParams.hpp>
+#include <popups/BedtimeReminderPopupRequestParams.hpp>
 #include <service-appmgr/Actions.hpp>
 #include <service-appmgr/messages/ActionRequest.hpp>
 
@@ -39,16 +39,12 @@ class AlarmDeactivated : public sys::DataMessage, public app::manager::actions::
 class BedtimeNotification : public sys::DataMessage, public app::manager::actions::ConvertibleToAction
 {
   public:
-    BedtimeNotification(Record record) : sys::DataMessage{MessageType::MessageTypeUninitialized}, eventRecord{record}
+    BedtimeNotification() : sys::DataMessage{MessageType::MessageTypeUninitialized}
     {}
 
     [[nodiscard]] auto toAction() const -> std::unique_ptr<app::manager::ActionRequest> override
     {
         return std::make_unique<app::manager::ActionRequest>(
-            sender,
-            app::manager::actions::ShowPopup,
-            std::make_unique<gui::BedtimeReminderPopupRequestParams>(eventRecord));
+            sender, app::manager::actions::ShowPopup, std::make_unique<gui::BedtimeReminderPopupRequestParams>());
     }
-
-    Record eventRecord;
 };
