+++ products/BellHybrid/services/time/tests/CMakeLists.txt
@@ -4,7 +4,6 @@ add_gtest_executable(
     SRCS
         test-BellAlarmOperations.cpp
     LIBS
-        service-time-alarm-operations-source
         module-sys
         bell::time
     INCLUDE
